restructure folders and create admin tab
