+++ loan-calculator/src/CSS/App.css
@@ -1,4 +1,3 @@
-
 .App-header {
   background-color: #580;
   height: 8vh;
