+++ loan-calculator/src/SCRIPTS/calculators.js
