+++ loan-calculator/src/components/Admin.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function Admin() {
+	return (<div><h1>This is the admin page</h1></div>)
+}
