+++ loan-calculator/src/components/AffordabilityCalculator.js
