+++ loan-calculator/src/components/CashBackCalculator.js
