+++ loan-calculator/src/components/UserRooms.js
@@ -1,0 +1,6 @@
+import React from 'react'
+
+
+export default function UserRooms() {
+	return(<div>This page will hold all of the user rooms to be joined by an admin/regions rep</div>)
+}
